Search opportunities page.
Step 2 of the disection of people and opportunities into to separate
pages rather than a single browse page.
