--- assets/js/backbone/apps/browse/templates/browse_search_tag.html
@@ -1,7 +1,0 @@
-<li data-id="<%- data.id %>">
-  <% /* escaped in browser_main_view */ %>
-  <%= format %>
-  <div class="pull-right">
-    <a href="#" class="search-tag-remove"><i class="fa fa-remove"></i></a>
-  </div>
-</li>
