--- assets/js/backbone/apps/browse/templates/profile_list_template.html
@@ -1,39 +1,0 @@
-<div class="container-fluid padding-top-none padding-bottom-none box" id="profile-list">
-  <table class="table">
-
-    <thead>
-    <tr>
-      <th width="20%">Name</th>
-      <th width="30%">Title</th>
-      <th width="25%" data-i18n="tag.Agency" ></th>
-      <th width="25%">Location</th>
-    </tr>
-    </thead>
-
-    <% _.each(people, function(person) { %>
-    <% if (!_.isUndefined(person.name)) { %>
-    <tr>
-      <td class="name">
-        <a href="/profile/<%= person.id %>">
-          <%= person.name %>
-        </a>
-      </td>
-      <td>
-        <%- person.title %>
-      </td>
-      <td>
-        <% if (!_.isUndefined(person.agency)) { %>
-        <%- person.agency.name %>
-        <% } %>
-      </td>
-      <td>
-        <% if (!_.isUndefined(person.location)) { %>
-        <%- person.location.name %>
-        <% } %>
-      </td>
-    </tr>
-    <% } %>
-    <% }); %>
-
-  </table>
-</div>
