--- assets/js/backbone/apps/browse/templates/profile_map_tooltip.html
@@ -1,10 +1,0 @@
-<b>
-  <%= city %>
-</b>
-<br>
-<%= count %>
-<% if (count === 1) { print("person:"); }
-   else if (count > 3) { print("people, including"); }
-   else { print("people:"); }
-%>
-<%- names %>
