--- assets/js/backbone/apps/browse/templates/project_list_item.html
@@ -1,46 +1,0 @@
-<li data-id="<%- item.id %>" class="box project-box">
-  <% if (item.coverId) { %>
-  <div class="project-background-image border-bottom" style="background-image: url(/api/upload/get/<%- item.coverId %>)">
-  <% } else { %>
-  <div>
-  <% } %>
-  </div>
-  <div class="project-title">
-    <a href="/projects/<%- item.id %>" class="link-backbone c1">
-      <%- item.title %>
-    </a>
-  </div>
-  <div class="project-list-description">
-    <% /* Escaped by marked() */ %>
-    <%= item.descriptionHtml %>
-  </div>
-
-  <div class="project-list-bottom-wrapper border-top">
-    <div class="project-list-counts">
-      <div class="project-list-counts-col">
-        <div class="project-list-count c2">
-          <%- item.taskCount %>
-        </div>
-        <div class="project-list-count-text c2">
-          opp<% if (item.taskCount != 1) { %>s<% } %>
-        </div>
-      </div>
-      <div class="project-list-counts-col">
-        <div class="project-list-count c2">
-          <%- item.commentCount %>
-        </div>
-        <div class="project-list-count-text c2">
-          comment<% if (item.commentCount != 1) { %>s<% } %>
-        </div>
-      </div>
-      <div class="project-list-counts-col">
-        <div class="project-list-count c2">
-          <%- item.ownerCount %>
-        </div>
-        <div class="project-list-count-text c2  ">
-          <% if (item.ownerCount != 1) { %>people<% } else { %>person<% } %>
-        </div>
-      </div>
-    </div>
-  </div>
-</li>
