+++ assets/js/backbone/apps/tasks/list/controllers/task_list_controller.js
@@ -4,81 +4,22 @@ var Backbone = require('backbone');
 
 var Bootstrap = require('bootstrap');
 var TasksCollection = require('../../../../entities/tasks/tasks_collection');
-var TaskCollectionView = require('../views/task_collection_view');
-// var TaskFormView = require('../../new/views/task_form_view');
-var ModalWizardComponent = require('../../../../components/modal_wizard');
 var TaskModel = require('../../../../entities/tasks/task_model');
+var TaskListView = require('../views/task_list_view');
 
-TaskList = Backbone.View.extend({
-
-  el: '#task-list-wrapper',
-
+TaskController = Backbone.View.extend({
   events: {
-    'click .add-task' : 'add',
-    'click .show-task': 'show',
-    'click .task'     : 'show',
-    'click .wizard'   : 'wizard',
+    'click .add-opportunity' : 'add',
   },
 
-  initialize: function (settings) {
-    this.options = _.extend(settings, this.defaults);
+  initialize: function (options) {
     var self = this;
-
-    this.initializeTaskCollectionInstance();
-    this.initializeTaskModelInstance();
-    this.initializeListeners();
-    this.requestTasksCollectionData();
-
-    this.collection.on('tasks:render', function () {
-      self.requestTasksCollectionData();
-    });
-  },
-
-  initializeListeners: function () {
-    var self = this;
-    this.listenTo(this.taskModel, 'task:tags:save:success', function () {
-      self.initializeTaskModelInstance();
-      self.requestTasksCollectionData();
-    });
-  },
-
-  initializeTaskModelInstance: function () {
-    this.taskModel = new TaskModel();
-  },
-
-  initializeTaskCollectionInstance: function () {
-    if (this.collection) {
-      this.collection.initialize();
-    } else {
-      this.collection = new TasksCollection();
-    }
-  },
-
-  requestTasksCollectionData: function () {
-    var self = this;
-
-    this.collection.fetch({
-      url: '/api/task/findAllByProjectId/' + parseInt(this.options.projectId),
-      success: function (collection) {
-        self.tasks = collection;
-        self.renderTaskCollectionView();
-      },
-    });
-  },
-
-  renderTaskCollectionView: function () {
-    var self = this;
-
-    this.listenTo(this.collection, 'task:save:success', function (model) {
-      self.requestTasksCollectionData();
-    });
-
-    if (this.taskCollectionView) this.taskCollectionView.cleanup();
-    this.taskCollectionView = new TaskCollectionView({
-      el: '#task-list-wrapper',
-      onRender: true,
-      collection: self.tasks,
-    });
+    self.options = options;
+    new TaskListView({
+      collection: new TasksCollection(),
+      el: self.el,
+      queryParams: self.options.queryParams,
+    }).render();
   },
 
   add: function (e) {
@@ -96,12 +37,9 @@ TaskList = Backbone.View.extend({
   },
 
   cleanup: function () {
-    if (this.taskFormView) this.taskFormView.cleanup();
-    if (this.modalWizardComponent) this.modalWizardComponent.cleanup();
-    if (this.taskCollectionView) this.taskCollectionView.cleanup();
     removeView(this);
   },
 
 });
 
-module.exports = TaskList;
+module.exports = TaskController;
