--- assets/js/backbone/apps/tasks/list/templates/task_collection_view_template.html
@@ -1,39 +1,0 @@
-<div class="box-pad-lr border-bottom">
-  <h2>
-    Available Opportunities
-    <% if (user) { %>
-    <a href="#addTask" class="btn btn-c0 btn-sm file-add add-task" data-toggle="modal" data-backdrop="static">Add Opportunity</a>
-    <% } %>
-  </h2>
-</div>
-
-<div class="task-scroll">
-  <% if (tasks.length == 0) { %>
-  <div class="task border-left" data-id="null">
-    <div class="task-title">
-      No current opportunities.
-    </div>
-  </div>
-  <% } %>
-  <% _.each(tasks, function(task) { %>
-  <div class="task border-left" data-projectid="<%- task.projectId %>" data-id="<%- task.id %>">
-    <div class="pull-right">
-      <strong><%- task.state %></strong>
-    </div>
-    <div class="task-title">
-      <a href="/tasks/<%- task.id %>" class="show-task" data-projectid="<%- task.projectId %>" data-id="<%- task.id %>"><%- task.title %></a>
-    </div>
-    <div class="task-description">
-      <% /* Escaped with marked() */ %>
-      <%= task.description %>
-    </div>
-    <div class="task-tags-row">
-      <ul class="tags task-tags">
-        <% _.each(task.tags, function (tag) { %>
-          <li><%- tag.name %></li>
-        <% }); %>
-      </ul>
-    </div>
-  </div>
-  <% }) %>
-</div>
