+++ package.json
@@ -146,6 +146,6 @@
   "bugs": "https://github.com/openopps/openopps-platform/issues",
   "license": "CC0-1.0",
   "engines": {
-    "node": "8.9.3"
+    "node": "8.9.x"
   }
 }
